[spotify] Restore libspotify as an alternative streaming backend
Can be enabled via configuration
