+++ src/input.h
@@ -16,8 +16,11 @@ enum input_types
   INPUT_TYPE_HTTP,
   INPUT_TYPE_PIPE,
   INPUT_TYPE_TIMER,
-#ifdef SPOTIFY
+#ifdef SPOTIFY_SPOTIFYC
   INPUT_TYPE_SPOTIFY,
+#endif
+#ifdef SPOTIFY_LIBSPOTIFY
+  INPUT_TYPE_LIBSPOTIFY,
 #endif
 };
