+++ src/library/spotify_webapi.c
@@ -410,11 +410,7 @@ token_get(const char *code, const char *redirect_uri, const char **err)
   keyval_clear(&kv);
 
   if (ret == 0)
-    {
-      request_user_info();
-      const char *errmsg;
-      spotify_login_token(spotify_credentials.user, (uint8_t *) spotify_credentials.access_token, strlen(spotify_credentials.access_token), &errmsg);
-    }
+    request_user_info();
 
   CHECK_ERR(L_SPOTIFY, pthread_mutex_unlock(&token_lck));
 
@@ -979,10 +975,9 @@ spotifywebapi_oauth_uri_get(const char *redirect_uri)
 
 /* Thread: httpd */
 int
-spotifywebapi_oauth_callback(struct evkeyvalq *param, const char *redirect_uri, char **errmsg)
+spotifywebapi_oauth_callback(struct evkeyvalq *param, const char *redirect_uri, const char **errmsg)
 {
   const char *code;
-  const char *err;
   int ret;
 
   *errmsg = NULL;
@@ -990,18 +985,19 @@ spotifywebapi_oauth_callback(struct evkeyvalq *param, const char *redirect_uri, 
   code = evhttp_find_header(param, "code");
   if (!code)
     {
-      *errmsg = safe_asprintf("Error: Didn't receive a code from Spotify");
+      *errmsg = "Error: Didn't receive a code from Spotify";
       return -1;
     }
 
   DPRINTF(E_DBG, L_SPOTIFY, "Received OAuth code: %s\n", code);
 
-  ret = token_get(code, redirect_uri, &err);
+  ret = token_get(code, redirect_uri, errmsg);
   if (ret < 0)
-    {
-      *errmsg = safe_asprintf("Error: %s", err);
-      return -1;
-    }
+    return -1;
+
+  ret = spotify_login_token(spotify_credentials.user, (uint8_t *) spotify_credentials.access_token, strlen(spotify_credentials.access_token), errmsg);
+  if (ret < 0)
+    return -1;
 
   // Trigger scan after successful access to spotifywebapi
   spotifywebapi_fullrescan();
@@ -1475,6 +1471,9 @@ track_add(struct spotify_track *track, struct spotify_album *album, const char *
       free_mfi(&mfi, 1);
     }
 
+  // This is only required for the libspotify backend
+  spotify_uri_register(track->uri);
+
   if (album && album->uri)
     cache_artwork_ping(track->uri, album->mtime, 0);
   else
@@ -1714,7 +1713,7 @@ scan_playlists(enum spotify_request_type request_type)
 }
 
 static void
-create_saved_tracks_playlist()
+create_saved_tracks_playlist(void)
 {
   struct playlist_info pli =
     {
@@ -1740,7 +1739,7 @@ create_saved_tracks_playlist()
  * Add or update playlist folder for all spotify playlists (if enabled in config)
  */
 static void
-create_base_playlist()
+create_base_playlist(void)
 {
   cfg_t *spotify_cfg;
   struct playlist_info pli =
@@ -1797,17 +1796,17 @@ scan(enum spotify_request_type request_type)
 
 /* Thread: library */
 static int
-initscan()
+initscan(void)
 {
   int ret;
 
   /* Refresh access token for the spotify webapi */
-  ret =  token_refresh();
+  ret = token_refresh();
   if (ret < 0)
     {
       DPRINTF(E_LOG, L_SPOTIFY, "Spotify webapi token refresh failed. "
-	"In order to use the web api, authorize the server to access "
-	"your saved tracks by visiting http://owntone.local:3689\n");
+	"In order to use Spotify, authorize the server to access your saved "
+	"tracks by visiting http://owntone.local:3689\n");
 
       db_spotify_purge();
 
@@ -1815,6 +1814,21 @@ initscan()
     }
 
   spotify_saved_plid = 0;
+
+  /*
+   * libspotify needs to be logged in before before scanning tracks from the web
+   * since scanned tracks need to be registered for playback
+   */
+  ret = spotify_relogin();
+  if (ret < 0)
+    {
+      DPRINTF(E_LOG, L_SPOTIFY, "libspotify-login failed. In order to use Spotify, "
+	"provide valid credentials for libspotify by visiting http://owntone.local:3689\n");
+
+      db_spotify_purge();
+
+      return 0;
+    }
 
   /*
    * Scan saved tracks from the web api
@@ -1826,7 +1840,7 @@ initscan()
 
 /* Thread: library */
 static int
-rescan()
+rescan(void)
 {
   scan(SPOTIFY_REQUEST_TYPE_RESCAN);
   return 0;
@@ -1834,7 +1848,7 @@ rescan()
 
 /* Thread: library */
 static int
-metarescan()
+metarescan(void)
 {
   scan(SPOTIFY_REQUEST_TYPE_METARESCAN);
   return 0;
@@ -1842,7 +1856,7 @@ metarescan()
 
 /* Thread: library */
 static int
-fullrescan()
+fullrescan(void)
 {
   db_spotify_purge();
   scan(SPOTIFY_REQUEST_TYPE_RESCAN);
@@ -2055,13 +2069,16 @@ spotifywebapi_init()
 {
   CHECK_ERR(L_SPOTIFY, mutex_init(&token_lck));
 
-  return 0;
+  // Required for libspotify backend
+  return spotify_init();
 }
 
 static void
 spotifywebapi_deinit()
 {
   CHECK_ERR(L_SPOTIFY, pthread_mutex_destroy(&token_lck));
+
+  spotify_deinit();
 
   free_credentials();
 }
