+++ src/library/spotify_webapi.h
@@ -45,7 +45,7 @@ struct spotifywebapi_access_token
 char *
 spotifywebapi_oauth_uri_get(const char *redirect_uri);
 int
-spotifywebapi_oauth_callback(struct evkeyvalq *param, const char *redirect_uri, char **errmsg);
+spotifywebapi_oauth_callback(struct evkeyvalq *param, const char *redirect_uri, const char **errmsg);
 
 void
 spotifywebapi_fullrescan(void);
