+++ src/misc.c
@@ -70,6 +70,12 @@ static char *buildopts[] =
 #else
     "Without Spotify",
 #endif
+#ifdef SPOTIFY_LIBRESPOTC
+    "librespot-c",
+#endif
+#ifdef SPOTIFY_LIBSPOTIFY
+    "libspotify",
+#endif
 #ifdef LASTFM
     "LastFM",
 #else
